Express Routes search & violation reports
- routes & controllers (logic) for searching & reporting violations
