+++ server/controller/violation_report-controller.js
@@ -1,0 +1,5 @@
+exports.get = (req, res) => {
+	res.render('reportViolations.hbs', {
+		pageTitle: 'Report Violations'
+  })
+};
